Current Weather API call

+++ app/src/main/java/app/itsyour/weathercrane/main/MainViewState.kt
@@ -1,4 +1,4 @@
 package app.itsyour.weathercrane.main
 
 sealed class MainViewState
-class WindSpeedUpdated(val windSpeed: Int) : MainViewState()
+class WindSpeedUpdated(val windSpeed: Double) : MainViewState()
